Redesign main cycle
Add WAITING phase to the main cycle, where Coordinator just waits for
robots to make a decision (this is a necessary change to add a support
for ensembles in future). Remove counting cycles on each robot
separately as it was confusing, and after these changes is redundant.
Make SpatialInput on robots' sensors carry the last action that was
performed by robot. It can inform robot about how his last action
ended up. Change robot classes according to changes in the main cycle.
Fix a bug in SimulationEngine.